+++ src/Pim/Bundle/DataGridBundle/EventListener/ConfigureGroupProductGridListener.php
@@ -13,7 +13,7 @@ use Pim\Bundle\DataGridBundle\Datagrid\Product\GroupColumnsConfigurator;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ConfigureGroupProductGridListener extends ConfigureFlexibleGridListener
+class ConfigureGroupProductGridListener extends ConfigureProductGridListener
 {
     /**
      * @param DatagridConfiguration $datagridConfig